pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "agentazall"
version = "1.0.17"
description = "Persistent memory and multi-agent communication — three transports (AgentTalk, Email, FTP), all open, all self-hostable"
readme = "README.md"
license = "AGPL-3.0-or-later"
requires-python = ">=3.10"
authors = [
    { name = "Gregor Koch" },
]
keywords = ["llm", "agent", "memory", "communication", "multi-agent", "agenttalk", "email", "ftp", "smtp", "imap", "rest-api"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "PyNaCl>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/cronos3k/AgentAZAll"
Repository = "https://github.com/cronos3k/AgentAZAll"
Issues = "https://github.com/cronos3k/AgentAZAll/issues"
"Live Demo" = "https://huggingface.co/spaces/cronos3k/AgentAZAll"
"PyPI" = "https://pypi.org/project/agentazall/"

[project.scripts]
agentazall = "agentazall.cli:main"

[project.optional-dependencies]
ftp = ["pyftpdlib>=1.5.9"]
ui = ["gradio>=4.0"]
dev = ["pytest>=7.0", "ruff>=0.4"]
all = ["pyftpdlib>=1.5.9", "gradio>=4.0", "pytest>=7.0", "ruff>=0.4"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
